Added tests for the rest of the methods

+++ oop/Calculator.cs
@@ -15,6 +15,6 @@ namespace oop
 
         public int Multiply(int a,int b) { return a * b; }
 
-        public double Divide(int a,int b) { return a / b; }
+        public double Divide(double a, double b) { return a / b; }
     }
 }
